IGNITE-24934 Extracted to seprate class.
