+++ modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/exp/IgniteSqlFunctions.java
@@ -17,7 +17,6 @@
 
 package org.apache.ignite.internal.sql.engine.exec.exp;
 
-import static org.apache.calcite.avatica.util.DateTimeUtils.dateStringToUnixDate;
 import static org.apache.calcite.runtime.SqlFunctions.charLength;
 import static org.apache.ignite.internal.lang.IgniteStringFormatter.format;
 import static org.apache.ignite.internal.sql.engine.prepare.IgniteSqlValidator.NUMERIC_FIELD_OVERFLOW_ERROR;
@@ -27,19 +26,13 @@ import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.math.RoundingMode;
 import java.time.LocalTime;
-import java.time.Year;
-import java.util.TimeZone;
 import java.util.UUID;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import org.apache.calcite.DataContext;
 import org.apache.calcite.DataContext.Variable;
 import org.apache.calcite.avatica.util.ByteString;
-import org.apache.calcite.avatica.util.DateTimeUtils;
 import org.apache.calcite.linq4j.function.NonDeterministic;
 import org.apache.calcite.runtime.SqlFunctions;
 import org.apache.calcite.sql.type.SqlTypeName;
-import org.apache.ignite.internal.lang.IgniteStringBuilder;
 import org.apache.ignite.internal.schema.SchemaUtils;
 import org.apache.ignite.internal.sql.engine.util.Commons;
 import org.apache.ignite.internal.sql.engine.util.IgniteMath;
@@ -60,14 +53,6 @@ public class IgniteSqlFunctions {
 
     private static final int DATE_MIN_INTERNAL = (int) TypeUtils.toInternal(SchemaUtils.DATE_MIN, NativeTypeSpec.DATE);
     private static final int DATE_MAX_INTERNAL = (int) TypeUtils.toInternal(SchemaUtils.DATE_MAX, NativeTypeSpec.DATE);
-
-    /** Regex for time, HH:MM:SS. */
-    private static final Pattern ISO_TIME_PATTERN =
-            Pattern.compile("^([0-2]\\d):[0-5]\\d:[0-5]\\d(\\.\\d*)*$");
-
-    /** Regex for date, YYYY-MM-DD. */
-    private static final Pattern ISO_DATE_PATTERN =
-            Pattern.compile("^(\\d{4})-([0]\\d|1[0-2])-([0-2]\\d|3[01])$");
 
     /**
      * Default constructor.
@@ -568,244 +553,6 @@ public class IgniteSqlFunctions {
     /** Returns the second argument and ignores the first. */
     public static Object consumeFirstArgument(Object args0, Object args1) {
         return args1;
-    }
-
-    /** Returns the timestamp value minus the offset of the specified timezone. */
-    public static Long subtractTimeZoneOffset(long timestamp, TimeZone timeZone) {
-        // A second offset calculation is required to handle DST transition period correctly.
-        int offset = timeZone.getOffset(timestamp - timeZone.getOffset(timestamp));
-
-        // After adjusting to UTC, you need to make sure that the value matches the allowed values.
-        return toTimestampLtzExact(timestamp - offset);
-    }
-
-    /**
-     * Helper for CAST({time} AS VARCHAR(n)).
-     *
-     * <p>Note: this method is a copy of the avatica {@link DateTimeUtils#unixTimestampToString(long, int)} method,
-     *          with the only difference being that it does not add trailing zeros.
-     */
-    public static String unixTimeToString(int time, int precision) {
-        IgniteStringBuilder buf = new IgniteStringBuilder(8 + (precision > 0 ? 1 + precision : 0));
-
-        unixTimeToString(buf, time, precision);
-
-        return buf.toString();
-    }
-
-    private static void unixTimeToString(IgniteStringBuilder buf, int time, int precision) {
-        int h = time / 3600000;
-        int time2 = time % 3600000;
-        int m = time2 / 60000;
-        int time3 = time2 % 60000;
-        int s = time3 / 1000;
-        int ms = time3 % 1000;
-
-        buf.app((char) ('0' + (h / 10) % 10))
-                .app((char) ('0' + h % 10))
-                .app(':')
-                .app((char) ('0' + (m / 10) % 10))
-                .app((char) ('0' + m % 10))
-                .app(':')
-                .app((char) ('0' + (s / 10) % 10))
-                .app((char) ('0' + s % 10));
-
-        if (precision == 0 || ms == 0) {
-            return;
-        }
-
-        buf.app('.');
-        do {
-            buf.app((char) ('0' + (ms / 100)));
-
-            ms = ms % 100;
-            ms = ms * 10;
-            --precision;
-        } while (ms > 0 && precision > 0);
-    }
-
-    /**
-     * Helper for CAST({timestamp} AS VARCHAR(n)).
-     *
-     * <p>Note: this method is a copy of the avatica {@link DateTimeUtils#unixTimestampToString(long, int)} method,
-     *          with the only difference being that it does not add trailing zeros.
-     */
-    public static String unixTimestampToString(long timestamp, int precision) {
-        IgniteStringBuilder buf = new IgniteStringBuilder(17);
-        int date = (int) (timestamp / DateTimeUtils.MILLIS_PER_DAY);
-        int time = (int) (timestamp % DateTimeUtils.MILLIS_PER_DAY);
-
-        if (time < 0) {
-            --date;
-
-            time += (int) DateTimeUtils.MILLIS_PER_DAY;
-        }
-
-        buf.app(DateTimeUtils.unixDateToString(date)).app(' ');
-
-        unixTimeToString(buf, time, precision);
-
-        return buf.toString();
-    }
-
-    /**
-     * Converts timestamp string into unix timestamp.
-     *
-     * <p>Note: this method is a copy of the avatica {@link DateTimeUtils#timestampStringToUnixDate(String)} method,
-     *          with the only difference that result is truncated to milliseconds without rounding.
-     *
-     * @param s Timestamp string.
-     * @return Timestamp.
-     */
-    public static long timestampStringToUnixDate(String s) {
-        try {
-            final long d;
-            final long t;
-            s = s.trim();
-            int space = s.indexOf(' ');
-            if (space >= 0) {
-                String datePart = s.substring(0, space);
-                validateDate(datePart, s);
-                d = dateStringToUnixDate(datePart);
-
-                String timePart = s.substring(space + 1);
-                validateTime(timePart, s);
-                t = timeStringToUnixDate(timePart);
-            } else {
-                validateDate(s, s);
-                d = dateStringToUnixDate(s);
-                t = 0;
-            }
-            return d * DateTimeUtils.MILLIS_PER_DAY + t;
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException(e.getMessage());
-        }
-    }
-
-    /**
-     * Converts time string into unix time.
-     *
-     * <p>Note: this method is a copy of the avatica {@link DateTimeUtils#timeStringToUnixDate(String)} method,
-     *          with the only difference that result is truncated to milliseconds without rounding.
-     *
-     * @param v Time string.
-     * @return Timestamp.
-     */
-    public static int timeStringToUnixDate(String v) {
-        final int colon1 = v.indexOf(':');
-        int hour;
-        int minute;
-        int second;
-        int milli = 0;
-        if (colon1 < 0) {
-            hour = Integer.parseInt(v.trim());
-            minute = 0;
-            second = 0;
-        } else {
-            hour = Integer.parseInt(v.substring(0, colon1).trim());
-            final int colon2 = v.indexOf(':', colon1 + 1);
-            if (colon2 < 0) {
-                minute = Integer.parseInt(v.substring(colon1 + 1).trim());
-                second = 0;
-            } else {
-                minute = Integer.parseInt(v.substring(colon1 + 1, colon2).trim());
-                int dot = v.indexOf('.', colon2);
-                if (dot < 0) {
-                    second = Integer.parseInt(v.substring(colon2 + 1).trim());
-                } else {
-                    second = Integer.parseInt(v.substring(colon2 + 1, dot).trim());
-                    String fraction = v.substring(dot + 1).trim();
-                    int multiplier = 100;
-                    for (int i = 0; i < Math.min(fraction.length(), 3); i++) {
-                        char c = fraction.charAt(i);
-                        int x = c < '0' || c > '9' ? 0 : (c - '0');
-                        milli += multiplier * x;
-                        multiplier /= 10;
-                    }
-                }
-            }
-        }
-        return hour * (int) DateTimeUtils.MILLIS_PER_HOUR
-                + minute * (int) DateTimeUtils.MILLIS_PER_MINUTE
-                + second * (int) DateTimeUtils.MILLIS_PER_SECOND
-                + milli;
-    }
-
-    private static void validateTime(String time, String full) {
-        Matcher matcher = ISO_TIME_PATTERN.matcher(time);
-        if (matcher.find()) {
-            int hour = Integer.parseInt(matcher.group(1));
-            if (hour > 23) {
-                throw fieldOutOfRange("HOUR", full);
-            }
-        } else {
-            throw invalidType("TIME", full);
-        }
-    }
-
-    private static void validateDate(String s, String full) {
-        Matcher matcher = ISO_DATE_PATTERN.matcher(s);
-        if (matcher.find()) {
-            int year = Integer.parseInt(matcher.group(1));
-            int month = Integer.parseInt(matcher.group(2));
-            int day = Integer.parseInt(matcher.group(3));
-            checkLegalDate(year, month, day, full);
-        } else {
-            throw invalidType("DATE", full);
-        }
-    }
-
-    /** Check that the combination year, month, date forms a legal date. */
-    private static void checkLegalDate(int year, int month, int day, String full) {
-        if (day > daysInMonth(year, month)) {
-            throw fieldOutOfRange("DAY", full);
-        }
-        if (month < 1 || month > 12) {
-            throw fieldOutOfRange("MONTH", full);
-        }
-        if (year <= 0) {
-            // Year 0 is not really a legal value.
-            throw fieldOutOfRange("YEAR", full);
-        }
-    }
-
-    /** Returns the number of days in a month in the proleptic Gregorian calendar
-     * used by ISO-8601.
-     *
-     * <p>"Proleptic" means that we apply the calendar to dates before the
-     * Gregorian calendar was invented (in 1582). Thus, years 0 and 1200 are
-     * considered leap years, and 1500 is not. */
-    private static int daysInMonth(int year, int month) {
-        switch (month) {
-            case 9:
-            case 4:
-            case 6:
-            case 11:
-                // Thirty days hath September,
-                // April, June, and November,
-                return 30;
-
-            default:
-                // All the rest have thirty-one,
-                return 31;
-
-            case 2:
-                // Except February, twenty-eight days clear,
-                // And twenty-nine in each leap year.
-                return Year.isLeap(year) ? 29 : 28;
-        }
-    }
-
-    private static IllegalArgumentException fieldOutOfRange(String field,
-            String full) {
-        return new IllegalArgumentException("Value of " + field
-                + " field is out of range in '" + full + "'");
-    }
-
-    private static IllegalArgumentException invalidType(String type,
-            String full) {
-        return new IllegalArgumentException("Invalid " + type + " value, '"
-                + full + "'");
     }
 
     private static @Nullable Object leastOrGreatest(boolean least, Object arg0, Object arg1) {
